+++ gazon-service/src/main/java/com/gitlab/model/ReviewImage.java
@@ -1,0 +1,30 @@
+package com.gitlab.model;
+
+
+import lombok.*;
+
+import javax.persistence.*;
+
+@Entity
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@EqualsAndHashCode(exclude = "review")
+@Table(name = "review_image")
+public class ReviewImage {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne
+    @JoinColumn(name = "review_id", referencedColumnName = "id")
+    private Review review;
+
+    @Column(name = "name")
+    private String name;
+
+    @Column(name = "data")
+    private byte[] data;
+}