+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - #6 - added pagination for Example
 - #7 - entities Product and ProductImage were created and integrated
 - #10 - entity SelectedProduct was created, but not fully integrated
+- #9 - entities Review and ReviewImage were created, but not fully integrated
 
 ## [1.0.0]